Music Controls - Embed Button Function
