+++ .gitignore
@@ -11,4 +11,4 @@ dist
 .env
 application.yaml
 *.bat
-
+prisma/migrations/
